add missing benchmarks
