+++ evmd/tests/integration/x_vm_test.go
@@ -33,27 +33,3 @@ func TestVmAnteTestSuite(t *testing.T) {
 func TestIterateContracts(t *testing.T) {
 	vm.TestIterateContracts(t, CreateEvmd)
 }
-
-func BenchmarkApplyTransaction(b *testing.B) {
-	vm.BenchmarkApplyTransaction(b, CreateEvmd)
-}
-
-func BenchmarkApplyTransactionWithLegacyTx(b *testing.B) {
-	vm.BenchmarkApplyTransactionWithLegacyTx(b, CreateEvmd)
-}
-
-func BenchmarkApplyTransactionWithDynamicFeeTx(b *testing.B) {
-	vm.BenchmarkApplyTransactionWithDynamicFeeTx(b, CreateEvmd)
-}
-
-func BenchmarkApplyMessage(b *testing.B) {
-	vm.BenchmarkApplyMessage(b, CreateEvmd)
-}
-
-func BenchmarkApplyMessageWithLegacyTx(b *testing.B) {
-	vm.BenchmarkApplyMessageWithLegacyTx(b, CreateEvmd)
-}
-
-func BenchmarkApplyMessageWithDynamicFeeTx(b *testing.B) {
-	vm.BenchmarkApplyMessageWithDynamicFeeTx(b, CreateEvmd)
-}
