+++ tests/integration/x/vm/state_transition_benchmark.go
@@ -2,7 +2,6 @@ package vm
 
 import (
 	"errors"
-	"github.com/cosmos/evm/testutil/integration/evm/network"
 	"math/big"
 	"testing"
 
@@ -12,6 +11,7 @@ import (
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/stretchr/testify/require"
 
+	"github.com/cosmos/evm/testutil/integration/evm/network"
 	utiltx "github.com/cosmos/evm/testutil/tx"
 	evmtypes "github.com/cosmos/evm/x/vm/types"
